requirements.txt
# General
# ------------------------------------------------------------------------------
gunicorn==20.1.0
phonenumbers==8.13.26
requests==2.31.0

# Django
# ------------------------------------------------------------------------------
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-crontab==0.7.1
django-phonenumber-field==7.2.0
django-filter==23.4
django-choices==2.0.0

# Restframework
# ------------------------------------------------------------------------------
djangorestframework==3.14.0

# celery
# ------------------------------------------------------------------------------
celery==5.3.6

# s3
# ------------------------------------------------------------------------------
boto3==1.34.36
botocore==1.34.50

# DB
# ------------------------------------------------------------------------------
psycopg2-binary==2.9.9


# In requirements.txt, add these lines:
langchain
langchain-google-genai
python-dotenv
requests
pydantic

# Add these lines to requirements.txt
celery
redis

dj-database-url

